Make some tests to run in serial mode.
